update: updated dtelib to version2 and included 2 new notebooks
